The first verison of generated.pictures !
I'll still surprised, though. I was looking for a quick
domain name to put my gallery under, and when I typed
generated.pictures, the domain was available !

So I took it, since it's exactly what my site is about.

The gallery code is currently extremely simple. I'll have
to improve the overall look of the gallery, but I'd like
to keep the code pretty simple.

The code include a PNG chunk parser, in order to get
the metadata back, since in 2022 reading PNG metadata is
very hard... This era still surprises me.

Signed-off-by: Voyage <voyage@miouyouyou.fr>